# platform-monitor/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Monitor System API
  description: |
    Comprehensive monitoring system API for managing clients, collecting metrics, and providing real-time monitoring capabilities.
    
    ## Features
    - Client registration and management
    - Real-time metrics collection
    - Historical data querying
    - Data export capabilities
    - WebSocket real-time communication
    - Certificate-based authentication
    
    ## Authentication
    The API uses JWT tokens for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <jwt-token>
    ```
  version: 2.0.0
  contact:
    name: Monitor System Support
    email: support@monitor.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://monitor.yourdomain.com
    description: Production server
  - url: https://monitor-staging.yourdomain.com
    description: Staging server
  - url: http://localhost:8080
    description: Development server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/auth/ask-code:
    get:
      tags:
        - Authentication
      summary: Request email verification code
      description: Send verification code to email for registration or password reset
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
          example: user@example.com
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [register, reset]
          example: reset
      responses:
        '200':
          description: Verification code sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/reset-confirm:
    post:
      tags:
        - Authentication
      summary: Confirm password reset
      description: Verify email and code for password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - code
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                code:
                  type: string
                  example: "123456"
      responses:
        '200':
          description: Reset confirmation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid code or email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/reset-password:
    post:
      tags:
        - Authentication
      summary: Reset password
      description: Reset user password with verification code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - code
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                code:
                  type: string
                  example: "123456"
                password:
                  type: string
                  minLength: 8
                  example: newpassword123
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid request or weak password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/logout:
    delete:
      tags:
        - Authentication
      summary: User logout
      description: Invalidate current JWT token
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Monitor Management Endpoints
...
  /api/monitor/register:
    get:
      tags:
        - Monitor Management
      summary: Generate registration token
      description: Generate a token for client registration (admin only)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Registration token generated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: string
                        example: "reg_token_abc123def456"
        '403':
          description: Forbidden - admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/exports/sync:
    post:
      tags:
        - Data Export
      summary: Export data synchronously
      description: Export data and return immediately (for small datasets)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
      responses:
        '200':
          description: Data exported successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ExportResult'
  /api/v1/exports/async:
    post:
      tags:
        - Data Export
      summary: Export data asynchronously
      description: Start async export job for large datasets
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
      responses:
        '202':
          description: Export job started
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ExportResult'
  /api/v1/exports/{exportId}/status:
    get:
      tags:
        - Data Export
      summary: Get export status
      description: Check status of async export job
      security:
        - BearerAuth: []
      parameters:
        - name: exportId
          in: path
          required: true
          schema:
            type: string
          example: export-123456
      responses:
        '200':
          description: Export status retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ExportResult'
  /api/v1/exports/{exportId}:
    delete:
      tags:
        - Data Export
      summary: Cancel export
      description: Cancel running export job
      security:
        - BearerAuth: []
      parameters:
        - name: exportId
          in: path
          required: true
          schema:
            type: string
          example: export-123456
      responses:
        '200':
          description: Export cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/exports:
    get:
      tags:
        - Data Export
      summary: List exports
      description: Get list of export jobs with pagination
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
          example: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
          example: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, PROCESSING, COMPLETED, FAILED, CANCELLED]
          example: COMPLETED
      responses:
        '200':
          description: Export list retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ExportResult'

  # Client Registration Endpoints
  /monitor/register:
    get:
      tags:
        - Client Registration
      summary: Register client
      description: Register monitoring client with registration token
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Client registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Invalid registration token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /monitor/detail:
    post:
      tags:
        - Client Registration
      summary: Update client details
      description: Update client system information
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientDetail'
      responses:
        '200':
          description: Client details updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /monitor/runtime:
    post:
      tags:
        - Client Registration
      summary: Update runtime details
      description: Submit current runtime metrics
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuntimeDetail'
      responses:
        '200':
          description: Runtime details updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "操作成功"
        data:
          type: object
        timestamp:
          type: string
          format: date-time
          example: "2024-10-17T10:30:00Z"
        requestId:
          type: string
          example: "req-123456"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "错误描述"
        error:
          type: object
          properties:
            code:
              type: string
              example: "ERROR_CODE"
            details:
              type: string
              example: "详细错误信息"
            timestamp:
              type: string
              format: date-time
              example: "2024-10-17T10:30:00Z"
            requestId:
              type: string
              example: "req-123456"
...
    ClientDetail:
      type: object
      required:
        - osName
        - osVersion
        - ip
        - cpuName
        - cpuCore
        - memory
        - disk
      properties:
        osName:
          type: string
          example: "Ubuntu"
        osVersion:
          type: string
          example: "20.04.3 LTS"
        ip:
          type: string
          example: "192.168.1.100"
        cpuName:
          type: string
          example: "Intel Core i7-8700K"
        cpuCore:
          type: integer
          example: 8
        memory:
          type: integer
          example: 16384
        disk:
          type: integer
          example: 512000
    RuntimeDetail:
      type: object
      required:
        - timestamp
        - cpuUsage
        - memoryUsage
        - diskUsage
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2024-10-17T10:30:00Z"
        cpuUsage:
          type: number
          format: double
          example: 45.2
        memoryUsage:
          type: number
          format: double
          example: 68.5
        diskUsage:
          type: number
          format: double
          example: 78.3
        networkUpload:
          type: integer
          example: 1024
        networkDownload:
          type: integer
          example: 2048
        diskRead:
          type: integer
          example: 512
        diskWrite:
          type: integer
          example: 256
...
    ExportRequest:
      type: object
      required:
        - clientIds
        - startTime
        - endTime
        - format
      properties:
        clientIds:
          type: array
          items:
            type: string
          example: ["client-001", "client-002"]
        startTime:
          type: string
          format: date-time
          example: "2024-10-17T00:00:00Z"
        endTime:
          type: string
          format: date-time
          example: "2024-10-17T23:59:59Z"
        metrics:
          type: array
          items:
            type: string
          example: ["cpu.usage", "memory.usage"]
        format:
          type: string
          enum: [CSV, JSON, EXCEL]
          example: "CSV"
        aggregation:
          type: string
          enum: [RAW, HOURLY, DAILY]
          example: "HOURLY"
        filters:
          type: object
          additionalProperties:
            type: string
          example:
            environment: "production"
        options:
          type: object
          properties:
            includeHeaders:
              type: boolean
              example: true
            compression:
              type: string
              enum: [NONE, GZIP, ZIP]
              example: "GZIP"
    ExportResult:
      type: object
      properties:
        exportId:
          type: string
          example: "export-123456"
        status:
          type: string
          enum: [PENDING, PROCESSING, COMPLETED, FAILED, CANCELLED]
          example: "COMPLETED"
        progress:
          type: number
          format: double
          example: 100.0
        downloadUrl:
          type: string
          example: "https://monitor.yourdomain.com/api/v1/exports/export-123456/download"
        fileSize:
          type: integer
          example: 1048576
        recordCount:
          type: integer
          example: 10000
        createdAt:
          type: string
          format: date-time
          example: "2024-10-17T10:30:00Z"
        completedAt:
          type: string
          format: date-time
          example: "2024-10-17T10:35:00Z"
        estimatedCompletionTime:
          type: string
          format: date-time
          example: "2024-10-17T10:35:00Z"
        error:
          type: string
          example: "Export failed due to timeout"
...